requirement.txt
sentence-transformers
uvicorn
ctransformers
langchain
python-box
streamlit
loadenv
google.generativeai